requirements.txt
python-docx==1.1.2
requests==2.31.0
google-generativeai==0.8.3
openai>=1.12.0
httpx>=0.25.0